requirements.txt
beautifulsoup4==4.12.3
chainlit==1.1.402
python-dotenv==1.0.1
pandas==2.2.2
qdrant-client==1.11.0
requests==2.32.3
llama-index==0.11.1
llama-index-llms-openai==0.2.0
llama-index-vector-stores-qdrant==0.3.0